Code/Raspberry_Pi/requirements.txt
mfrc522
openpyxl
pigpio
pillow
types-pillow
pygame
pre-commit
